Do not execute defs and the main entrypoint during compilation
Background:
During compilation each statement is compiled and executed. The execution is required for the macro extension. Since macros can also use other definitions like `def` or `defn`, we can not decide before the compilation with one we need. Therefore everything is compiled.

Solution:
The `*compile-mode*` flag is usually set to `false`. However, during compilation it is set to `true`. We can use this flag as a gate for code that should not be executed during compilation.